refactor: test 파일 패키지 구조를 main 과 일치시킴
