+++ src/test/kotlin/racingcar/domain/CarTest.kt
@@ -1,10 +1,9 @@
-package racingcar
+package racingcar.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatExceptionOfType
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
-import racingcar.domain.Car
 
 class CarTest {
     @ParameterizedTest
