+++ src/test/kotlin/racingcar/domain/CarsTest.kt
@@ -1,10 +1,8 @@
-package racingcar
+package racingcar.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
-import racingcar.domain.Cars
-import racingcar.domain.IRandomNumberGenerator
 
 class CarsTest {
     @ParameterizedTest
